:recycle: refactor(github-search): migrate to ESM next.config + shared CSP (#559)
Convert next.config.js to .mjs and consume @howardism/security-headers
with a connect-src override admitting https://api.github.com for Apollo
Client. Inline securityHeaders array removed.
