+++ apps/github-search/package.json
@@ -19,6 +19,7 @@
   "dependencies": {
     "@apollo/client": "^4.1.6",
     "@howardism/components-common": "workspace:*",
+    "@howardism/security-headers": "workspace:*",
     "@howardism/ui": "workspace:*",
     "graphql": "^16.8.1",
     "graphql-tag": "^2.12.6",
